requirements.txt
Django==3.2.9
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2-binary==2.9.2
djoser==2.1.0 
djangorestframework-simplejwt==4.8.0
pytest==6.2.5
pytest-django==4.4.0